[#168340845]  Validating user input  (#40)
* Chore(Server) Refactoring to Module Based

Make routes modules
Refactor codes
Improving with ES6 standards
Removing repetitive functions

[starts #168314608]

* Refactored code with Tests

Refaactoring Tests
Writing unit tests
Chnaged bug in UI

[start ch-refactoring-#168314608]
"

* chore(Server) Refactoring-codes

Refactor Codes
Increase test coverage

[finishes #168314608]

* Updating Coveralls Badge

Updating Coveralls Badge

[finish #168314608]

* Chore(Server) Removing duplicate File

* chore(server) Removed no longer used function due to refactoring

Remove reapeting function which are no longer in use
[finishes #168314608]

* chore(Server): Validating input

Validating input
Increasing tests

[finish #168340845]

+++ server/controllers/sessionController.js
@@ -7,11 +7,13 @@ const sessionController = {
   createSessions: (req, res) => {
     const data = sessionModel.createSession(req.body, req.userData);
     const status = 201;
-    const message = 'undefined';
+
+    const message = 'Session request was sent, waiting mentor approval';
     return Helper.handleSuccess(res, status, message, data);
   },
   // Getting All data
-  getSessions: (req, res) => { // Middleware for Checking Role Needed    
+  getSessions: (req, res) => {
+
     const roles = ['user', 'mentor', 'admin'];
     const userRole = Helper.getRole(req.userData.role);
 
@@ -21,7 +23,9 @@ const sessionController = {
       return Helper.handleError(res, status, error);
     }
     const allSessions = sessionModel.findSessions();
-    const userSessions = Helper.filterObjectByProp(allSessions, userRole, req.userData.userId)
+
+    const userSessions = Helper.filterObjectByProp(allSessions, userRole, req.userData.userId);
+
     if (req.userData.role === 'admin') {
       const status = 200;
       const message = 'undefined';
@@ -35,7 +39,8 @@ const sessionController = {
   },
 
 
-  // Acceppt Session
+  // Accept Session
+
   acceptSession: (req, res) => {
     sessionModel.changeStatus('Accepted', req.params.sessionId, res);
   },
@@ -45,15 +50,20 @@ const sessionController = {
   },
 
   createReview: (req, res) => {
-    const data = reviewModel.createReview(req.body, req.params.sessionId, req.userData);
-    const status = 200;
-    const message = 'undefined';
-    return Helper.handleSuccess(res, status, message, data);
+    return reviewModel.createReview(req.body, req.params.sessionId, req.userData, res);
+
   },
 
   // Deleting Data
   deleteReview: (req, res) => {
-    const result = reviewModel.deleteReview(req.params.sessionId, res);
+
+    if (req.userData.role === 'admin') {
+      return reviewModel.deleteReview(req.params.sessionId, res);
+    }
+    const status = 401;
+    const error = 'You are not allowed to change status';
+    return Helper.handleError(res, status, error);
+
   },
 
 };
